services/aegis-iso20022-api/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.38.0
pydantic==2.12.3
lxml==6.0.2
PyYAML==6.0.1
xmlschema>=2.0,<3
httpx==0.27.2
python-multipart>=0.0.6
ulid-py==1.1.0
